Working on animation structures - employing state design pattern

+++ src/main/java/stickman/view/background/BackgroundItem.java
@@ -5,5 +5,5 @@ import javafx.scene.layout.Pane;
 public interface BackgroundItem {
     void draw(Pane windowPane);
 
-    void update(int viewportOffset);
+    void update(double viewportOffset);
 }
